+++ packages/cds-annotation-parser/eslint.config.mjs
@@ -1,17 +1,7 @@
 import base from '../../eslint.config.mjs';
 
-const __dirname = import.meta.dirname;
-
 export default [
     ...base,
-    {
-        languageOptions: {
-            parserOptions: {
-                tsconfigRootDir: __dirname,
-                project: './tsconfig.eslint.json',
-            }
-        }
-    },
     {
         'rules': {
             '@typescript-eslint/naming-convention': [