+++ packages/adp-tooling/src/writer/project-utils.ts
@@ -1,10 +1,9 @@
-import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
+import { dirname, join } from 'node:path';
 import { readFileSync } from 'node:fs';
 import type { Editor } from 'mem-fs-editor';
 
-const __dirname = dirname(fileURLToPath(import.meta.url));
-
+import { getTemplatePath } from '../templates.js';
 import type { CloudApp, AdpWriterConfig, TypesConfig, CfAdpWriterConfig, DescriptorVariant } from '../types.js';
 import {
     enhanceUI5DeployYaml,
@@ -19,6 +18,8 @@ import {
 
 import type { Package } from '@sap-ux/project-access';
 import { UI5Config, UI5_DEFAULT, getEsmTypesVersion, getTypesPackage, getTypesVersion } from '@sap-ux/ui5-config';
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
 
 /**
  * Retrieves the package name and version from the package.json file located two levels up the directory tree.
@@ -227,8 +228,7 @@ export async function writeCfTemplates(
     config: CfAdpWriterConfig,
     fs: Editor
 ): Promise<void> {
-    const baseTmplPath = join(__dirname, '../../templates');
-    const templatePath = config.options?.templatePathOverwrite ?? baseTmplPath;
+    const templatePath = config.options?.templatePathOverwrite ?? getTemplatePath();
     const { app, project, options } = config;
 
     fs.copyTpl(