+++ packages/abap-deploy-config-sub-generator/CHANGELOG.md
@@ -20,6 +20,8 @@
 
 ## 1.0.17
 
+*Released: 2026-06-16T21:19:36Z*
+
 ### Patch Changes
 
 #### Workspace Updates
@@ -28,6 +30,8 @@
 
 ## 1.0.16
 
+*Released: 2026-06-14T10:40:09Z*
+
 ### Patch Changes
 
 - Updated dependencies [adae40d]
@@ -35,12 +39,16 @@
 
 ## 1.0.15
 
+*Released: 2026-06-12T10:49:08Z*
+
 ### Patch Changes
 
 - Updated dependencies [6e4a71a]
     - @sap-ux/adp-tooling@1.0.15
 
 ## 1.0.14
+
+*Released: 2026-06-12T06:53:23Z*
 
 ### Patch Changes
 
@@ -55,12 +63,16 @@
 
 ## 1.0.13
 
+*Released: 2026-06-11T13:37:16Z*
+
 ### Patch Changes
 
 - Updated dependencies [e66a8a9]
     - @sap-ux/adp-tooling@1.0.13
 
 ## 1.0.12
+
+*Released: 2026-06-11T10:54:17Z*
 
 ### Patch Changes
 
@@ -70,6 +82,8 @@
 - @sap-ux/adp-tooling@1.0.12
 
 ## 1.0.11
+
+*Released: 2026-06-10T09:57:42Z*
 
 ### Patch Changes
 
@@ -85,6 +99,8 @@
 
 ## 1.0.10
 
+*Released: 2026-06-09T14:35:01Z*
+
 ### Patch Changes
 
 - Updated dependencies [0fa8305]
@@ -98,6 +114,8 @@
 
 ## 1.0.9
 
+*Released: 2026-06-09T13:18:16Z*
+
 ### Patch Changes
 
 - Updated dependencies [a328e14]
@@ -107,6 +125,8 @@
     - @sap-ux/adp-tooling@1.0.9
 
 ## 1.0.8
+
+*Released: 2026-06-04T13:54:21Z*
 
 ### Patch Changes
 
@@ -121,12 +141,16 @@
 
 ## 1.0.7
 
+*Released: 2026-06-04T12:10:05Z*
+
 ### Patch Changes
 
 - Updated dependencies [fb84c0e]
     - @sap-ux/adp-tooling@1.0.7
 
 ## 1.0.6
+
+*Released: 2026-06-04T10:19:37Z*
 
 ### Patch Changes
 
@@ -140,6 +164,8 @@
     - @sap-ux/deploy-config-generator-shared@1.0.5
 
 ## 1.0.5
+
+*Released: 2026-06-03T14:58:37Z*
 
 ### Patch Changes
 
@@ -159,6 +185,8 @@
 
 ## 1.0.4
 
+*Released: 2026-06-03T13:52:44Z*
+
 ### Patch Changes
 
 - Updated dependencies [9580241]
@@ -173,6 +201,8 @@
 
 ## 1.0.3
 
+*Released: 2026-06-02T11:35:17Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-inquirer@2.0.2
@@ -182,12 +212,16 @@
 
 ## 1.0.2
 
+*Released: 2026-06-02T08:56:31Z*
+
 ### Patch Changes
 
 - Updated dependencies [a8e4cf0]
     - @sap-ux/adp-tooling@1.0.2
 
 ## 1.0.1
+
+*Released: 2026-06-01T15:15:26Z*
 
 ### Patch Changes
 
@@ -201,6 +235,8 @@
     - @sap-ux/deploy-config-generator-shared@1.0.1
 
 ## 1.0.0
+
+*Released: 2026-05-30T20:54:07Z*
 
 ### Major Changes
 
@@ -250,6 +286,8 @@
 
 ## 0.4.11
 
+*Released: 2026-05-29T12:50:34Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-inquirer@1.9.7
@@ -258,12 +296,16 @@
 
 ## 0.4.10
 
+*Released: 2026-05-29T06:59:27Z*
+
 ### Patch Changes
 
 - Updated dependencies [9a980a9]
     - @sap-ux/adp-tooling@0.19.10
 
 ## 0.4.9
+
+*Released: 2026-05-27T11:39:21Z*
 
 ### Patch Changes
 
@@ -273,6 +315,8 @@
 - @sap-ux/fiori-generator-shared@0.15.6
 
 ## 0.4.8
+
+*Released: 2026-05-27T10:42:47Z*
 
 ### Patch Changes
 
@@ -284,6 +328,8 @@
 - @sap-ux/abap-deploy-config-writer@0.4.4
 
 ## 0.4.7
+
+*Released: 2026-05-26T16:40:21Z*
 
 ### Patch Changes
 
@@ -299,12 +345,16 @@
 
 ## 0.4.6
 
+*Released: 2026-05-22T13:30:05Z*
+
 ### Patch Changes
 
 - Updated dependencies [758b0d4]
     - @sap-ux/adp-tooling@0.19.6
 
 ## 0.4.5
+
+*Released: 2026-05-21T16:21:11Z*
 
 ### Patch Changes
 
@@ -320,12 +370,16 @@
 
 ## 0.4.4
 
+*Released: 2026-05-21T14:58:44Z*
+
 ### Patch Changes
 
 - Updated dependencies [dab1aa2]
     - @sap-ux/adp-tooling@0.19.4
 
 ## 0.4.3
+
+*Released: 2026-05-19T15:16:46Z*
 
 ### Patch Changes
 
@@ -340,12 +394,16 @@
 
 ## 0.4.2
 
+*Released: 2026-05-18T08:15:14Z*
+
 ### Patch Changes
 
 - Updated dependencies [342c544]
     - @sap-ux/adp-tooling@0.19.2
 
 ## 0.4.1
+
+*Released: 2026-05-15T13:12:06Z*
 
 ### Patch Changes
 
@@ -358,6 +416,8 @@
     - @sap-ux/deploy-config-generator-shared@0.2.1
 
 ## 0.4.0
+
+*Released: 2026-05-15T08:12:20Z*
 
 ### Minor Changes
 
@@ -380,6 +440,8 @@
 
 ## 0.3.37
 
+*Released: 2026-05-14T21:28:41Z*
+
 ### Patch Changes
 
 - @sap-ux/adp-tooling@0.18.134
@@ -388,12 +450,16 @@
 
 ## 0.3.36
 
+*Released: 2026-05-14T14:16:50Z*
+
 ### Patch Changes
 
 - Updated dependencies [8c4185a]
     - @sap-ux/adp-tooling@0.18.133
 
 ## 0.3.35
+
+*Released: 2026-05-14T11:45:51Z*
 
 ### Patch Changes
 
@@ -413,6 +479,8 @@
 
 ## 0.3.34
 
+*Released: 2026-05-13T09:36:59Z*
+
 ### Patch Changes
 
 - Updated dependencies [21abda3]
@@ -426,6 +494,8 @@
 
 ## 0.3.33
 
+*Released: 2026-05-12T18:00:39Z*
+
 ### Patch Changes
 
 - Updated dependencies [9360ea5]
@@ -435,6 +505,8 @@
     - @sap-ux/adp-tooling@0.18.130
 
 ## 0.3.32
+
+*Released: 2026-05-06T23:02:00Z*
 
 ### Patch Changes
 
@@ -453,6 +525,8 @@
 
 ## 0.3.31
 
+*Released: 2026-05-01T15:46:09Z*
+
 ### Patch Changes
 
 - Updated dependencies [b2ffc7e]
@@ -460,12 +534,16 @@
 
 ## 0.3.30
 
+*Released: 2026-04-30T19:47:20Z*
+
 ### Patch Changes
 
 - Updated dependencies [fcaa70c]
     - @sap-ux/adp-tooling@0.18.127
 
 ## 0.3.29
+
+*Released: 2026-04-30T14:23:24Z*
 
 ### Patch Changes
 
@@ -480,11 +558,15 @@
 
 ## 0.3.28
 
+*Released: 2026-04-30T13:10:33Z*
+
 ### Patch Changes
 
 - @sap-ux/adp-tooling@0.18.125
 
 ## 0.3.27
+
+*Released: 2026-04-29T15:24:37Z*
 
 ### Patch Changes
 
@@ -500,6 +582,8 @@
 
 ## 0.3.26
 
+*Released: 2026-04-27T19:47:46Z*
+
 ### Patch Changes
 
 - Updated dependencies [1d60871]
@@ -513,12 +597,16 @@
 
 ## 0.3.25
 
+*Released: 2026-04-27T07:30:24Z*
+
 ### Patch Changes
 
 - Updated dependencies [10847a1]
     - @sap-ux/adp-tooling@0.18.122
 
 ## 0.3.24
+
+*Released: 2026-04-23T12:54:21Z*
 
 ### Patch Changes
 
@@ -533,6 +621,8 @@
 
 ## 0.3.23
 
+*Released: 2026-04-23T06:48:55Z*
+
 ### Patch Changes
 
 - Updated dependencies [237371b]
@@ -545,6 +635,8 @@
 
 ## 0.3.22
 
+*Released: 2026-04-22T12:38:46Z*
+
 ### Patch Changes
 
 - @sap-ux/adp-tooling@0.18.119
@@ -552,6 +644,8 @@
 - @sap-ux/deploy-config-generator-shared@0.1.118
 
 ## 0.3.21
+
+*Released: 2026-04-15T11:53:17Z*
 
 ### Patch Changes
 
@@ -563,12 +657,16 @@
 
 ## 0.3.20
 
+*Released: 2026-04-15T08:11:32Z*
+
 ### Patch Changes
 
 - Updated dependencies [8fb08a2]
     - @sap-ux/adp-tooling@0.18.117
 
 ## 0.3.19
+
+*Released: 2026-04-14T20:26:28Z*
 
 ### Patch Changes
 
@@ -582,6 +680,8 @@
     - @sap-ux/abap-deploy-config-writer@0.3.10
 
 ## 0.3.18
+
+*Released: 2026-04-14T12:35:35Z*
 
 ### Patch Changes
 
@@ -598,6 +698,8 @@
 
 ## 0.3.17
 
+*Released: 2026-04-14T11:39:16Z*
+
 ### Patch Changes
 
 - Updated dependencies [497317c]
@@ -605,12 +707,16 @@
 
 ## 0.3.16
 
+*Released: 2026-04-09T07:00:16Z*
+
 ### Patch Changes
 
 - Updated dependencies [7a8613b]
     - @sap-ux/adp-tooling@0.18.113
 
 ## 0.3.15
+
+*Released: 2026-04-08T13:10:18Z*
 
 ### Patch Changes
 
@@ -629,6 +735,8 @@
 
 ## 0.3.14
 
+*Released: 2026-04-07T11:09:34Z*
+
 ### Patch Changes
 
 - Updated dependencies [1b10e9f]
@@ -636,12 +744,16 @@
 
 ## 0.3.13
 
+*Released: 2026-04-01T14:51:40Z*
+
 ### Patch Changes
 
 - Updated dependencies [6b74074]
     - @sap-ux/adp-tooling@0.18.110
 
 ## 0.3.12
+
+*Released: 2026-04-01T11:49:37Z*
 
 ### Patch Changes
 
@@ -656,12 +768,16 @@
 
 ## 0.3.11
 
+*Released: 2026-04-01T06:34:51Z*
+
 ### Patch Changes
 
 - Updated dependencies [68b5523]
     - @sap-ux/adp-tooling@0.18.108
 
 ## 0.3.10
+
+*Released: 2026-03-30T22:24:11Z*
 
 ### Patch Changes
 
@@ -687,6 +803,8 @@
 
 ## 0.3.9
 
+*Released: 2026-03-27T15:37:24Z*
+
 ### Patch Changes
 
 - @sap-ux/fiori-generator-shared@0.13.93
@@ -695,6 +813,8 @@
 - @sap-ux/adp-tooling@0.18.106
 
 ## 0.3.8
+
+*Released: 2026-03-27T11:58:49Z*
 
 ### Patch Changes
 
@@ -709,6 +829,8 @@
 
 ## 0.3.7
 
+*Released: 2026-03-26T20:06:10Z*
+
 ### Patch Changes
 
 - Updated dependencies [b66e827]
@@ -722,6 +844,8 @@
 
 ## 0.3.6
 
+*Released: 2026-03-26T16:48:32Z*
+
 ### Patch Changes
 
 - Updated dependencies [96a689b]
@@ -729,12 +853,16 @@
 
 ## 0.3.5
 
+*Released: 2026-03-26T15:15:10Z*
+
 ### Patch Changes
 
 - Updated dependencies [3dcd3f7]
     - @sap-ux/adp-tooling@0.18.102
 
 ## 0.3.4
+
+*Released: 2026-03-26T12:07:04Z*
 
 ### Patch Changes
 
@@ -762,6 +890,8 @@
 
 ## 0.3.3
 
+*Released: 2026-03-25T12:56:41Z*
+
 ### Patch Changes
 
 - Updated dependencies [f384ace]
@@ -775,6 +905,8 @@
 
 ## 0.3.2
 
+*Released: 2026-03-23T18:25:40Z*
+
 ### Patch Changes
 
 - Updated dependencies [c0e05ab]
@@ -786,6 +918,8 @@
     - @sap-ux/abap-deploy-config-writer@0.3.1
 
 ## 0.3.1
+
+*Released: 2026-03-23T10:51:10Z*
 
 ### Patch Changes